refactor translation page to pull out components and service code
